third_party/blink/web_tests/http/tests/security/window-named-proto-expected.txt

CONSOLE ERROR: Uncaught SecurityError: Failed to read a named property '__proto__' from 'Window': Blocked a frame with origin "null" from accessing a frame with origin "http://localhost:8080".  The frame requesting access has a protocol of "data", the frame being accessed has a protocol of "http". Protocols must match.
